Split out the helper functions, axis & legend components, groupedbar & line into separate files
